+++ lib/src/data/metrics/science.dart
@@ -1,0 +1,5 @@
+/// Metrics reported by the science subsystem. 
+/// 
+/// These metrics represent analysis of dirt samples extracted by the science subsystem. They need
+/// to not only be recorded but logged as well so the science team can generate charts out of it.
+// class ScienceMetrics extends Metrics<ScienceMessage> { }
